Add build tags for coverage
