+++ pkg/server/no_cover.go
@@ -1,0 +1,7 @@
+//go:build !cover
+
+package server
+
+import "context"
+
+func writeCoverage(ctx context.Context) {}
