+++ pkg/server/server.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"runtime/coverage"
 	"runtime/debug"
 	"strconv"
 	"strings"
@@ -95,26 +94,6 @@ func startOnAPIServerReady(ctx context.Context, config *Config) {
 	case <-config.ControlConfig.Runtime.APIServerReady:
 		if err := runControllers(ctx, config); err != nil {
 			logrus.Fatalf("failed to start controllers: %v", err)
-		}
-	}
-}
-
-// writeCoverage checks if GOCOVERDIR is set on startup and writes coverage files to that directory
-// every 20 seconds. This is done to ensure that the coverage files are written even if the process is killed.
-func writeCoverage(ctx context.Context) {
-	if k, ok := os.LookupEnv("GOCOVERDIR"); ok {
-		for {
-			select {
-			case <-ctx.Done():
-				if err := coverage.WriteCountersDir(k); err != nil {
-					logrus.Warn(err)
-				}
-				return
-			case <-time.After(20 * time.Second):
-				if err := coverage.WriteCountersDir(k); err != nil {
-					logrus.Warn(err)
-				}
-			}
 		}
 	}
 }
